KRACOEUS-7296: Prop Dev action page.  Integrated all changes from pulled upstream code and PR comments.

+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/core/ProposalDevelopmentSubmitController.java
@@ -6,22 +6,39 @@ import javax.servlet.http.HttpServletResponse;
 import org.kuali.coeus.propdev.impl.questionnaire.ProposalDevelopmentQuestionnaireHelper;
 import org.kuali.coeus.propdev.impl.s2s.question.ProposalDevelopmentS2sQuestionnaireHelper;
 import org.kuali.coeus.propdev.impl.state.ProposalState;
+import org.kuali.coeus.sys.framework.validation.AuditHelper;
+import org.kuali.rice.krad.util.GlobalVariables;
 import org.kuali.rice.krad.web.controller.MethodAccessible;
+import org.kuali.rice.krad.web.form.DialogResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
+@SuppressWarnings("deprecation")
 @Controller
 public class ProposalDevelopmentSubmitController extends
 		ProposalDevelopmentControllerBase {
 
+    @Autowired
+    @Qualifier("auditHelper")
+    private AuditHelper auditHelper;
+
+	
     @RequestMapping(value = "/proposalDevelopment", params = "methodToCall=deleteProposal")
     public ModelAndView deleteProposal(@ModelAttribute("KualiForm") ProposalDevelopmentDocumentForm form, BindingResult result,
             HttpServletRequest request, HttpServletResponse response) throws Exception {
-    	getProposalDevelopmentService().deleteProposal(form.getProposalDevelopmentDocument());
-        return getRefreshControllerService().refresh(form);
+        DialogResponse disapproveConfirm = form.getDialogResponse("PropDev-SubmitPage-ConfirmDelete");
+        if (disapproveConfirm == null) {
+            return getModelAndViewService().showDialog("PropDev-SubmitPage-ConfirmDelete", true, form);
+        }
+
+        getProposalDevelopmentService().deleteProposal(form.getProposalDevelopmentDocument());
+        return getNavigationControllerService().returnToHub(form);
+
     }
     @RequestMapping(value = "/proposalDevelopment", params="methodToCall=saveAndExit")
     public  ModelAndView saveAndExit(@ModelAttribute("KualiForm") ProposalDevelopmentDocumentForm form, BindingResult result, HttpServletRequest request, HttpServletResponse response)throws Exception{	
@@ -30,14 +47,21 @@ public class ProposalDevelopmentSubmitController extends
 	}
     @RequestMapping(value = "/proposalDevelopment", params="methodToCall=submitForReview")
     public  ModelAndView submitForReview(@ModelAttribute("KualiForm") ProposalDevelopmentDocumentForm form, BindingResult result, HttpServletRequest request, HttpServletResponse response)throws Exception {	
-    	form.getDevelopmentProposal().setSubmitFlag(true);
-    	return getRefreshControllerService().refresh(form);
+
+        AuditHelper.ValidationState state = getAuditHelper().isValidSubmission(form, true);
+        if (state != AuditHelper.ValidationState.ERROR){
+        	form.getDevelopmentProposal().setSubmitFlag(true);
+    		return getTransactionalDocumentControllerService().route(form);
+    	}else{
+    		GlobalVariables.getMessageMap().clearErrorMessages(); 
+    		return getModelAndViewService().showDialog("PropDev-DataValidationSection", true, form);
+    	}
    } 
    @RequestMapping(value = "/proposalDevelopment", params = "methodToCall=cancelProposal")
     public ModelAndView cancelProposal(@ModelAttribute("KualiForm") ProposalDevelopmentDocumentForm form, BindingResult result,
             HttpServletRequest request, HttpServletResponse response) throws Exception {
 	   form.getDevelopmentProposal().setProposalStateTypeCode(ProposalState.CANCELED);
-	   return getRefreshControllerService().refresh(form);
+	   return getTransactionalDocumentControllerService().cancel(form);
     }
     @MethodAccessible
     @RequestMapping(value = "/proposalDevelopment", params={"methodToCall=navigate", "actionParameters[navigateToPageId]=PropDev-SubmitPage"}) 
@@ -45,5 +69,13 @@ public class ProposalDevelopmentSubmitController extends
         ((ProposalDevelopmentViewHelperServiceImpl) form.getViewHelperService()).populateCreditSplits(form);
         ((ProposalDevelopmentViewHelperServiceImpl) form.getViewHelperService()).populateQuestionnaires(form);
         return getNavigationControllerService().navigate(form);
+   }
+  
+    public AuditHelper getAuditHelper() {
+        return auditHelper;
+    }
+
+    public void setAuditHelper(AuditHelper auditHelper) {
+        this.auditHelper = auditHelper;
     }
 }
